Fix a typo in ConvertToTensorShapeProto declaration.
PiperOrigin-RevId: 266509672

+++ tensorflow/compiler/mlir/tensorflow/utils/convert_tensor.h
@@ -42,7 +42,7 @@ StatusOr<mlir::ElementsAttr> ConvertTensor(const Tensor& input_tensor,
                                            mlir::Builder* builder);
 
 // Converts a shape from MLIR to an TensorFlow tensor shape proto.
-Status ConvertToTensorShapeProto(llvm::ArrayRef<int> shape,
+Status ConvertToTensorShapeProto(llvm::ArrayRef<int64_t> shape,
                                  TensorShapeProto* output_shape);
 
 // Converts an MLIR elements attribute to an TensorFlow tensor proto.
